989: Improved proxy handling logic
