+++ connectors/http/http-base/src/main/java/io/camunda/connector/http/base/client/apache/ProxyHandler.java
@@ -4,6 +4,7 @@ import io.camunda.connector.api.error.ConnectorInputException;
 import java.net.*;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import org.apache.commons.lang3.ArrayUtils;
@@ -30,78 +31,111 @@ public class ProxyHandler {
       String nonProxyHosts,
       boolean sourceIsSystemProperties) {}
 
-  private Map<String, ProxyDetails> map = new HashMap<>();
+  private static final List<String> PROTOCOLS = List.of("http", "https");
+  private Map<String, ProxyDetails> proxyConfigForProtocols = new HashMap<>();
+  private Map<String, CredentialsProvider> credentialsProvidersForProtocols = new HashMap<>();
   private static final Logger LOGGER = LoggerFactory.getLogger(ProxyHandler.class);
 
   public ProxyHandler() {
-    try {
-      setFromEnvVars("http");
-      setFromEnvVars("https");
-      setFromSystemProperties("http");
-      setFromSystemProperties("https");
-    } catch (NumberFormatException e) {
-      throw new ConnectorInputException("Invalid proxy port", e);
-    }
+    this.proxyConfigForProtocols = loadProxyConfig();
+    this.credentialsProvidersForProtocols = initializeCredentialsProviders();
   }
 
-  private void setFromEnvVars(String protocol) throws NumberFormatException {
-    if (isValidProxyFromEnvVars(protocol)) {
-      String host = System.getenv("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_HOST");
-      int port =
-          Integer.parseInt(System.getenv("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_PORT"));
-      String user =
-          System.getenv().getOrDefault("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_USER", "");
-      char[] password =
-          System.getenv()
-              .getOrDefault("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_PASSWORD", "")
-              .toCharArray();
-      String nonProxyHosts =
-          System.getenv("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_NON_PROXY_HOSTS");
-      map.put(
-          protocol, new ProxyDetails(protocol, host, port, user, password, nonProxyHosts, false));
-      LOGGER.debug("Proxy settings applied from environment variables for {}", protocol);
+  private Map<String, ProxyDetails> loadProxyConfig() {
+    Map<String, ProxyDetails> config = new HashMap<>();
+    for (String protocol : PROTOCOLS) {
+      Optional<ProxyDetails> details =
+          getConfigFromSystemProperties(protocol).or(() -> getConfigFromEnvVars(protocol));
+      details.ifPresent(d -> config.put(protocol, d));
     }
+    return config;
   }
 
-  private void setFromSystemProperties(String protocol) throws NumberFormatException {
-    if (isValidProxyFromSystemProperties(protocol)) {
-      String host = System.getProperty(protocol + ".proxyHost");
-      int port = Integer.parseInt(System.getProperty(protocol + ".proxyPort"));
-      String user = System.getProperty(protocol + ".proxyUser");
-      char[] password =
-          Optional.ofNullable(System.getProperty(protocol + ".proxyPassword"))
-              .map(String::toCharArray)
-              .orElseGet(() -> new char[0]);
-      String nonProxyHosts = System.getProperty(protocol + ".nonProxyHosts");
-      map.put(
-          protocol, new ProxyDetails(protocol, host, port, user, password, nonProxyHosts, true));
-      LOGGER.debug("Proxy settings applied from system properties for {}", protocol);
+  private Map<String, CredentialsProvider> initializeCredentialsProviders() {
+    Map<String, CredentialsProvider> providers = new HashMap<>();
+    for (Map.Entry<String, ProxyDetails> entry : proxyConfigForProtocols.entrySet()) {
+      ProxyDetails p = entry.getValue();
+      if (StringUtils.isNotBlank(p.user()) && ArrayUtils.isNotEmpty(p.password())) {
+        BasicCredentialsProvider provider = new BasicCredentialsProvider();
+        provider.setCredentials(
+            new AuthScope(p.host(), p.port()),
+            new UsernamePasswordCredentials(p.user(), p.password()));
+        providers.put(entry.getKey(), provider);
+      }
     }
+    return providers;
+  }
+
+  private Optional<ProxyDetails> getConfigFromEnvVars(String protocol) {
+    if (StringUtils.isNotBlank(System.getenv("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_HOST"))
+        && StringUtils.isNotBlank(
+            System.getenv("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_PORT"))) {
+      try {
+        return Optional.of(
+            new ProxyDetails(
+                protocol,
+                System.getenv("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_HOST"),
+                Integer.parseInt(
+                    System.getenv("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_PORT")),
+                System.getenv()
+                    .getOrDefault("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_USER", ""),
+                System.getenv()
+                    .getOrDefault("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_PASSWORD", "")
+                    .toCharArray(),
+                System.getenv("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_NON_PROXY_HOSTS"),
+                false));
+      } catch (NumberFormatException e) {
+        throw new ConnectorInputException("Invalid proxy port in environment variables", e);
+      }
+    }
+    return Optional.empty();
+  }
+
+  private Optional<ProxyDetails> getConfigFromSystemProperties(String protocol) {
+    if (StringUtils.isNotBlank(System.getProperty(protocol + ".proxyHost"))
+        && StringUtils.isNotBlank(System.getProperty(protocol + ".proxyPort"))) {
+      try {
+        return Optional.of(
+            new ProxyDetails(
+                protocol,
+                System.getProperty(protocol + ".proxyHost"),
+                Integer.parseInt(System.getProperty(protocol + ".proxyPort")),
+                System.getProperty(protocol + ".proxyUser"),
+                Optional.ofNullable(System.getProperty(protocol + ".proxyPassword"))
+                    .map(String::toCharArray)
+                    .orElseGet(() -> new char[0]),
+                System.getProperty(protocol + ".nonProxyHosts"),
+                true));
+      } catch (NumberFormatException e) {
+        throw new ConnectorInputException(
+            "Invalid proxy port in system properties", e); // TODO add protocol
+      }
+    }
+    return Optional.empty();
   }
 
   public CredentialsProvider getCredentialsProvider(String protocol) {
-    BasicCredentialsProvider provider = new BasicCredentialsProvider();
-    ProxyDetails p = map.get(protocol);
-    if (p != null
-        && StringUtils.isNotBlank(p.host())
-        && StringUtils.isNotBlank(p.user())
-        && ArrayUtils.isNotEmpty(p.password())) {
-      provider.setCredentials(
-          new AuthScope(p.host(), p.port()),
-          new UsernamePasswordCredentials(p.user(), p.password()));
-    }
-    return provider;
+    return credentialsProvidersForProtocols.getOrDefault(protocol, new BasicCredentialsProvider());
   }
 
   public HttpHost getProxyHost(String protocol, String requestUrl) {
-    ProxyDetails p = map.get(protocol);
-    return p == null || doesTargetMatchNonProxy(protocol, requestUrl)
-        ? null
-        : new HttpHost(p.protocol(), p.host(), p.port());
+    ProxyDetails p = proxyConfigForProtocols.get(protocol);
+    if (p == null || doesTargetMatchNonProxy(protocol, requestUrl)) {
+      LOGGER.debug("No proxy used for request URL: {}", requestUrl);
+      return null;
+    }
+    LOGGER.debug(
+        "Using proxy for {} - Host: {}, Port: {}, Source: {}",
+        protocol,
+        p.host(),
+        p.port(),
+        p.sourceIsSystemProperties() ? "System Properties" : "Environment Variables");
+
+    return new HttpHost(p.protocol(), p.host(), p.port());
   }
 
   private boolean doesTargetMatchNonProxy(String protocol, String requestUri) {
-    ProxyDetails p = map.get(protocol);
+    ProxyDetails p = proxyConfigForProtocols.get(protocol);
     return (p != null
         && p.nonProxyHosts() != null
         && Arrays.stream(p.nonProxyHosts().split("\\|"))
@@ -111,25 +145,19 @@ public class ProxyHandler {
   }
 
   public HttpRoutePlanner getRoutePlanner(String protocol, HttpHost proxyHost) {
-    ProxyDetails p = map.get(protocol);
+    ProxyDetails p = proxyConfigForProtocols.get(protocol);
+
     if (p != null && p.sourceIsSystemProperties()) {
+      LOGGER.debug("Using system default route planner for protocol: {}", protocol);
       return new SystemDefaultRoutePlanner(
           DefaultSchemePortResolver.INSTANCE, ProxySelector.getDefault());
     } else if (proxyHost != null) {
+      LOGGER.debug(
+          "Using default proxy route planner for protocol: {} with proxy: {}", protocol, proxyHost);
       return new DefaultProxyRoutePlanner(proxyHost);
     }
+
+    LOGGER.debug("No proxy route planner used for protocol: {}", protocol);
     return null;
-  }
-
-  private static boolean isValidProxyFromEnvVars(String protocol) {
-    return StringUtils.isNotBlank(
-            System.getenv("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_HOST"))
-        && StringUtils.isNotBlank(
-            System.getenv("CONNECTOR_" + protocol.toUpperCase() + "_PROXY_PORT"));
-  }
-
-  private static boolean isValidProxyFromSystemProperties(String protocol) {
-    return (StringUtils.isNotBlank(System.getProperty(protocol + ".proxyHost"))
-        && StringUtils.isNotBlank(System.getProperty(protocol + ".proxyPort")));
   }
 }
