+++ connectors/http/http-base/src/test/resources/squid.conf
@@ -10,15 +10,12 @@ acl CONNECT method CONNECT
 auth_param basic program /usr/lib/squid/basic_ncsa_auth /etc/squid/passwords
 auth_param basic realm Squid Proxy
 
-# Allow unauthenticated users on /path
-acl no_auth_url url_regex -i path
-http_access allow no_auth_url
-
-# Define authenticated user
+acl protected_path url_regex -i /protected
 acl authenticated_user proxy_auth REQUIRED
 
-# Allow authenticated users everywhere
+http_access allow !protected_path
+http_access deny protected_path !authenticated_user
 http_access allow authenticated_user
 
-# Deny all other access
+# Deny everything else (optional, for safety)
 http_access deny all
